+++ R/googleAuthR_generator.R
@@ -275,7 +275,10 @@ doHttrRequest <- function(url,
   
   if(checkGoogleAPIError(req)){
     content <- httr::content(req, as = "text", type = "application/json",encoding = "UTF-8")
-    content <- jsonlite::fromJSON(content)
+    content <- jsonlite::fromJSON(content, 
+                                  simplifyVector = FALSE,
+                                  simplifyDataFrame = TRUE,
+                                  simplifyMatrix = TRUE)
     req$content <- content
   }